requirements.txt
llama-index==0.10.19
flask==3.0.2
chromadb==0.4.24
llama-index-vector-stores-chroma==0.1.6
semantic_router==0.0.27
semantic-router[local]
llama-index-readers-smart-pdf-loader==0.1.3
python-dotenv==1.0.1